// ssyweb/SSYNEW/src/app/models/userurl.model.ts
export class UserUrl {
    url: string;
    userTpe: string;
    userid: number;
    userMobileNo: string;
    benName: string;
    userTypeName: string;
    deptUserid: number;
    userName: string;
    location: string;
    subDiv: string;
    district: string;
    checkRegType: string;
    skipAdditionaldata: string;
    loginType: string;
    chkAddData: string;
    token: string;
    encData: string;
    serviceProvider: string;
    isDoubleVerification: boolean;
}